refactor(mobile): 역 출구 안내 컴포넌트 분리
